requirements.txt
Flask==3.0.3
flask-login==0.6.3
flask-wtf==1.2.1
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
WTForms==3.1.2
